pyproject.toml
[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[project]
name = "stockpredict"
version = "0.1.0"
description = "Multi-horizon US equity analysis platform"
requires-python = ">=3.11"
dependencies = [
  # Data sources
  "ib_async>=1.0.1",
  "yfinance>=0.2.50",
  "fredapi>=0.5.2",
  "finnhub-python>=2.4.20",
  # Data processing
  "pandas>=2.2",
  "numpy>=1.26",
  "pyarrow>=16",
  # ML
  "lightgbm>=4.3",
  "scikit-learn>=1.5",
  # Web
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sse-starlette>=2.0",
  # DB
  "sqlalchemy>=2.0",
  "aiosqlite>=0.20",
  "greenlet>=3.0",
  # AI
  "openai>=1.50",
  "anthropic>=0.40",
  # Config + utils
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "jinja2>=3.1",
  "pyyaml>=6.0",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
  "httpx>=0.27",
  "ruff",
  "mypy",
]

[project.scripts]
stockpredict = "stockpredict.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
